# Split metrics (cached) and tracing (uncached) label decoding