In tracing there is no concern with cardinality and one can have PID as a label. With caching enabled, this means that we grow the cache map indefinitely, leaking memory until OOM stops us. Let's disable caching for tracing instead.