@bobrik 2024-03-23
Perhaps an LRU with 1k entries per config is a better option here.